requirements.txt
streamlit>=1.31
python-dateutil>=2.8
pandas>=2.0
openpyxl>=3.1
supabase>=2.3
